improved naming and organization of files
