+++ src/components/mainview/productdetail/specification/Case.jsx
@@ -1,7 +1,7 @@
 import React, {Component} from "react";
 import {connect} from "react-redux";
 
-class CaseDetail extends Component {
+class Case extends Component {
 
     render() {
         return (
@@ -53,4 +53,4 @@ const mapStateToProps = state => {
 
 export default connect(mapStateToProps, {
 
-})(CaseDetail);
+})(Case);
