+++ src/components/mainview/productdetail/specification/CentralProcessingUnit.jsx
@@ -1,7 +1,7 @@
 import React, {Component} from "react";
 import {connect} from "react-redux";
 
-class CPUDetail extends Component {
+class CentralProcessingUnit extends Component {
 
     render() {
         return (
@@ -45,4 +45,4 @@ const mapStateToProps = state => {
 
 export default connect(mapStateToProps, {
 
-})(CPUDetail);
+})(CentralProcessingUnit);
