+++ src/components/mainview/productdetail/specification/Keyboard.jsx
@@ -1,7 +1,7 @@
 import React, {Component} from "react";
 import {connect} from "react-redux";
 
-class KeyboardDetail extends Component {
+class Keyboard extends Component {
 
     render() {
         return (
@@ -53,4 +53,4 @@ const mapStateToProps = state => {
 
 export default connect(mapStateToProps, {
 
-})(KeyboardDetail);
+})(Keyboard);
