+++ src/components/mainview/productdetail/specification/Mouse.jsx
@@ -1,7 +1,7 @@
 import React, {Component} from "react";
 import {connect} from "react-redux";
 
-class MouseDetail extends Component {
+class Mouse extends Component {
 
     render() {
         return (
@@ -49,4 +49,4 @@ const mapStateToProps = state => {
 
 export default connect(mapStateToProps, {
 
-})(MouseDetail);
+})(Mouse);
